Improvements in background Elasticsearch index update (#51)
* Queue background tasks according to Microsoft documentation.

* Check if Elasticsearch sync feature is enabled before adding items to task queue.

+++ aspnetcore/src/api/Controllers/UserProfileController.cs
@@ -23,16 +23,21 @@ namespace api.Controllers
         private readonly KeycloakAdminApiService _keycloakAdminApiService;
         private readonly ILogger<UserProfileController> _logger;
         private readonly IMemoryCache _cache;
-        private readonly BackgroundElasticsearchPersonUpdateQueue _backgroundElasticsearchPersonUpdateQueue;
+        private readonly IBackgroundTaskQueue _taskQueue;
 
-        public UserProfileController(ElasticsearchService elasticsearchService, UserProfileService userProfileService, KeycloakAdminApiService keycloakAdminApiService, ILogger<UserProfileController> logger, IMemoryCache memoryCache, BackgroundElasticsearchPersonUpdateQueue backgroundElasticsearchPersonUpdateQueue)
+        public UserProfileController(ElasticsearchService elasticsearchService,
+            UserProfileService userProfileService,
+            KeycloakAdminApiService keycloakAdminApiService,
+            ILogger<UserProfileController> logger,
+            IMemoryCache memoryCache,
+            IBackgroundTaskQueue taskQueue)
         {
             _userProfileService = userProfileService;
             _elasticsearchService = elasticsearchService;
             _keycloakAdminApiService = keycloakAdminApiService;
             _logger = logger;
             _cache = memoryCache;
-            _backgroundElasticsearchPersonUpdateQueue = backgroundElasticsearchPersonUpdateQueue;
+            _taskQueue = taskQueue;
         }
 
         /// <summary>
@@ -113,13 +118,16 @@ namespace api.Controllers
             _cache.Remove(orcidId);
 
             // Remove entry from Elasticsearch index in a background task.
-            _backgroundElasticsearchPersonUpdateQueue.QueueBackgroundWorkItem(async token =>
+            if (_elasticsearchService.IsElasticsearchSyncEnabled())
             {
-                _logger.LogInformation($"Background task for removing {orcidId} from Elasticsearch person index started at {DateTime.UtcNow}");
-                // Update Elasticsearch person index.
-                await _elasticsearchService.DeleteEntryFromElasticsearchPersonIndex(orcidId);
-                _logger.LogInformation($"Background task for removing {orcidId} from Elasticseach person index ended at {DateTime.UtcNow}");
-            });
+                await _taskQueue.QueueBackgroundWorkItemAsync(async token =>
+                {
+                    _logger.LogInformation($"Elasticsearch removal of {orcidId} started {DateTime.UtcNow}");
+                    // Update Elasticsearch person index.
+                    await _elasticsearchService.DeleteEntryFromElasticsearchPersonIndex(orcidId);
+                    _logger.LogInformation($"Elasticsearch removal of {orcidId} completed {DateTime.UtcNow}");
+                });
+            }
 
             // Delete profile data from database
             try
